requirements.txt
aiomysql==0.2.0
alembic==1.13.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.1
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
